compilation warning fixed)

+++ trunk/src/apps/rfc5769/rfc5769check.c
@@ -78,7 +78,7 @@ static int check_oauth(void) {
 
 	const char mac_key[33] = "ZksjpweoixXmvn67534m";
 	const size_t mac_key_length=strlen(mac_key);
-	const uint64_t token_timestamp = 92470300704768;
+	const uint64_t token_timestamp = (uint64_t)(92470300704768LL);
 	const uint32_t token_lifetime = 3600;
 
 	const char kid[33] = "2783466234";
